added code to make the editor functionaly
