+++ web/package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "antd": "^4.5.3",
+    "awesome-debounce-promise": "^2.1.0",
     "axios": "^0.19.2",
     "dotenv": "^8.2.0",
     "draft-js": "^0.11.6",
