+++ web/src/app/apis/notes.api.js
@@ -10,11 +10,31 @@ export const createNote = async (data) => {
   }
 };
 
+export const editNote = async (id, data) => {
+  try {
+    const resp = await Axios.put("/notes/" + id, data);
+    return resp.data;
+  } catch (error) {
+    console.log("Network error: ", error);
+    return error.response.data;
+  }
+};
+
 export const getAllNotes = async (limit, last_id) => {
   try {
     const url = last_id ? `/notes/${limit}/${last_id}/` : `/notes/${limit}/`;
 
     const resp = await Axios.get(url);
+    return resp.data;
+  } catch (error) {
+    console.log("Network error: ", error);
+    return error.response.data;
+  }
+};
+
+export const getNote = async (id) => {
+  try {
+    const resp = await Axios.get(`/notes/${id}`);
     return resp.data;
   } catch (error) {
     console.log("Network error: ", error);
